Remove paper component, move data specific forms to common, others
